+++ android/src/main/java/com/smileidentity/react/views/SmileIDBiometricKYCView.kt
@@ -2,8 +2,6 @@ package com.smileidentity.react.views
 
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.lifecycle.ViewModelStoreOwner
 import androidx.lifecycle.viewmodel.compose.LocalViewModelStoreOwner
 import com.facebook.react.bridge.ReactApplicationContext
 import com.smileidentity.SmileID
@@ -24,9 +22,9 @@ class SmileIDBiometricKYCView(context: ReactApplicationContext) : SmileIDView(co
       return
     }
     composeView.apply {
-      setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnDetachedFromWindow)
+      val customViewModelStoreOwner = CustomViewModelStoreOwner()
       setContent {
-        CompositionLocalProvider(LocalViewModelStoreOwner provides (context as ViewModelStoreOwner)) {
+        CompositionLocalProvider(LocalViewModelStoreOwner provides customViewModelStoreOwner) {
           SmileID.BiometricKYC(
             idInfo = idInfo!!,
             userId = userId ?: rememberSaveable { randomUserId() },