+++ ios/View/BaseSmileIDView.swift
@@ -4,18 +4,33 @@ import SwiftUI
 class BaseSmileIDView: UIView {
     typealias ContentView = AnyView
     var contentView : AnyView?
-
-    init(frame: CGRect,contentView:AnyView) {
+    private var _onResult: RCTBubblingEventBlock?
+    var product : SmileIDProductModel?
+    
+    @objc var onResult: RCTBubblingEventBlock? {
+        get {
+            return _onResult
+        }
+        set {
+            _onResult = newValue
+            if newValue != nil {
+                product?.onResult = newValue
+            }
+        }
+    }
+    
+    init(frame: CGRect,contentView:AnyView,product:SmileIDProductModel) {
         self.contentView = contentView
+        self.product = product
         super.init(frame: frame)
         commonInit()
     }
-
+    
     required init?(coder aDecoder: NSCoder) {
         super.init(coder: aDecoder)
         commonInit()
     }
-
+    
     private func commonInit() {
         // Perform initialization tasks here
         // For example, setup subviews, add constraints, configure appearance
@@ -28,8 +43,10 @@ class BaseSmileIDView: UIView {
         hostingView.bottomAnchor.constraint(equalTo: bottomAnchor).isActive = true
         hostingView.rightAnchor.constraint(equalTo: rightAnchor).isActive = true
     }
-
+    
     func getView(product:SmileIDProductModel) -> AnyView {
         fatalError("Must be implemented by subclass")
     }
+    
+    
 }