+++ ios/ViewManagers/SmileIDBaseViewManager.swift
@@ -4,15 +4,24 @@ import SwiftUI
 
 class SmileIDBaseViewManager: RCTViewManager, SmileIDUIViewDelegate {
     var product = SmileIDProductModel()
+    @objc var onResult: RCTDirectEventBlock?
     func getView() -> UIView {
         fatalError("Must be implemented by subclass")
     }
-
+    
     override func view() -> UIView! {
         getView()
     }
-
-    @objc override static func requiresMainQueueSetup() -> Bool {
-        true
+    
+    override func customBubblingEventTypes() -> [String]! {
+        return ["onResult"]
+    }
+    
+    @objc override func constantsToExport() -> [AnyHashable : Any]! {
+        return ["onResult": "onResult"]
+    }
+    
+    override static func requiresMainQueueSetup() -> Bool {
+        return true
     }
 }