+++ ios/ViewManagers/SmileIDConsentViewManager.m
@@ -2,5 +2,6 @@
 #import <React/RCTViewManager.h>
 
 @interface RCT_EXTERN_MODULE(SmileIDConsentViewManager, RCTViewManager)
+RCT_EXPORT_VIEW_PROPERTY(onResult, RCTBubblingEventBlock);
 RCT_EXTERN_METHOD(setParams:(nonnull NSNumber *)node params:(NSDictionary *)params)
 @end