+++ ios/ViewManagers/SmileIDConsentViewManager.swift
@@ -6,7 +6,7 @@ import SwiftUI
 @objc(SmileIDConsentViewManager)
 class SmileIDConsentViewManager: SmileIDBaseViewManager {
     override func getView() -> UIView {
-        BaseSmileIDView(frame: .zero, contentView: AnyView(SmileIDConsentView(product: self.product)))
+         BaseSmileIDView(frame: .zero, contentView: AnyView(SmileIDConsentView(product: self.product)),product:self.product)
     }
 
     @objc func setParams(_ node: NSNumber, params: NSDictionary) {
@@ -14,7 +14,7 @@ class SmileIDConsentViewManager: SmileIDBaseViewManager {
          */
         DispatchQueue.main.async {
             if let component = self.bridge.uiManager.view(forReactTag: node) as? BaseSmileIDView {
-                let onResult = params["onResult"] as? RCTDirectEventBlock
+                let onResult = params["onResult"] as? RCTBubblingEventBlock
                 guard let partnerIcon = params["partnerIcon"] as? String else {
                     onResult?(["error": SmileIDError.unknown("partnerIcon is required to run show consent screen")])
                     return
@@ -35,12 +35,12 @@ class SmileIDConsentViewManager: SmileIDBaseViewManager {
                     onResult?(["error": SmileIDError.unknown("partnerPrivacyPolicy must be a valid url")])
                     return
                 }
+                self.product.onResult = onResult
                 self.product.partnerIcon = partnerIcon
                 self.product.partnerName = partnerName
                 self.product.productName = productName
                 self.product.partnerPrivacyPolicy = partnerPrivacyPolicyUrl
                 self.product.showAttribution = params["showAttribution"] as? Bool ?? true
-                self.product.onResult = params["onResult"] as? RCTDirectEventBlock
             }
         }
     }