+++ ios/ViewManagers/SmileIDDocumentVerificationViewManager.m
@@ -3,4 +3,5 @@
 
 @interface RCT_EXTERN_MODULE(SmileIDDocumentVerificationViewManager, RCTViewManager)
 RCT_EXTERN_METHOD(setParams:(nonnull NSNumber *)node params:(NSDictionary *)params)
+RCT_EXPORT_VIEW_PROPERTY(onResult, RCTBubblingEventBlock);
 @end